+++ src/main/java/de/samply/email/EmailKeyValuesFactory.java
@@ -1,8 +1,10 @@
 package de.samply.email;
 
+import de.samply.app.ProjectManagerConst;
 import de.samply.bridgehead.BridgeheadConfiguration;
 import de.samply.db.repository.*;
 import de.samply.frontend.FrontendService;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -17,6 +19,8 @@ public class EmailKeyValuesFactory {
     private final ProjectDocumentRepository projectDocumentRepository;
     private final BridgeheadAdminUserRepository bridgeheadAdminUserRepository;
 
+    private final String researchEnvironmentUrl;
+
     public EmailKeyValuesFactory(FrontendService frontendService,
                                  EmailContext emailContext,
                                  ProjectBridgeheadRepository projectBridgeheadRepository,
@@ -24,7 +28,8 @@ public class EmailKeyValuesFactory {
                                  UserRepository userRepository,
                                  BridgeheadConfiguration bridgeheadConfiguration,
                                  ProjectDocumentRepository projectDocumentRepository,
-                                 BridgeheadAdminUserRepository bridgeheadAdminUserRepository) {
+                                 BridgeheadAdminUserRepository bridgeheadAdminUserRepository,
+                                 @Value(ProjectManagerConst.CODER_BASE_URL_SV) String researchEnvironmentUrl) {
         this.frontendService = frontendService;
         this.emailContext = emailContext;
         this.projectBridgeheadRepository = projectBridgeheadRepository;
@@ -33,11 +38,14 @@ public class EmailKeyValuesFactory {
         this.bridgeheadConfiguration = bridgeheadConfiguration;
         this.projectDocumentRepository = projectDocumentRepository;
         this.bridgeheadAdminUserRepository = bridgeheadAdminUserRepository;
+        this.researchEnvironmentUrl = researchEnvironmentUrl;
     }
 
     public EmailKeyValues newInstance() {
-        return new EmailKeyValues(frontendService, emailContext, projectBridgeheadRepository, projectRepository,
-                userRepository, bridgeheadConfiguration, projectDocumentRepository, bridgeheadAdminUserRepository);
+        return new EmailKeyValues(
+                frontendService, emailContext, projectBridgeheadRepository, projectRepository,
+                userRepository, bridgeheadConfiguration, projectDocumentRepository,
+                bridgeheadAdminUserRepository, researchEnvironmentUrl);
     }
 
 }